+++ clang-tools-extra/clang-tidy/add_new_check.py
@@ -333,7 +333,7 @@ def update_checks_list(clang_tidy_path):
   def process_doc(doc_file):
     check_name = doc_file.replace('.rst', '')
 
-    with open(os.path.join(docs_dir, doc_file), 'r') as doc:
+    with open(os.path.join(docs_dir, doc_file), 'r', encoding='utf-8') as doc:
       content = doc.read()
       match = re.search('.*:orphan:.*', content)